Update values.yaml
Fixes false positive on L40s with WPA errors. 

Return for check when BM shapes that are not explicitly listed should be the same for VMs and return an empty list. Previously it return a string stating that the shape was unsupported. That string was being resulted in the false positive. 

Conditional now returns an empty list, VMs specific conditional was removed as it is now redundent.